requirements.txt
python-dotenv
tqdm
psycopg2-binary==2.9.9
Faker==19.13.0
pandas==2.2.2
streamlit==1.38.0

